Zapisywanie autorstwa.

+++ ReferenceArchiver/Model/Authorship.cs
@@ -7,7 +7,7 @@ namespace ReferenceArchiver.Model
 {
     public class Authorship
     {
-        public int ArticleId { get; set; }
+        public long ArticleId { get; set; }
         public int Number { get; set; }
         public int AuthorId { get; set; }
         public string InstitutionId { get; set; }
